editor/libeditor/tests/test_bug858918.html

<!DOCTYPE html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=858918
-->
<title>Test for Bug 858918</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="/tests/SimpleTest/EventUtils.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=858918">Mozilla Bug 858918</a>
<span contenteditable style="display:block;min-height:1em"></span>
<script>
var span = document.querySelector("span");
getSelection().collapse(span, 0);
document.execCommand("inserthtml", false, "<div>doesn't go in span</div>");
is(span.innerHTML, "<div>doesn't go in span</div>");
</script>
